generalized a bit to permit people to define the public port and the root of the app

+++ DockerFiles/Symfony2/entrypoint.sh
@@ -27,7 +27,11 @@ EOS
     # we make sure to start fresh
     rm app/config/parameters.yml
     composer install --no-interaction
-    rm -rf app/cache/*
+    if [ -d app/cache ]; then
+        rm -rf app/cache/*
+    else
+        mkdir app/cache
+    fi
     php app/console assets:install --symlink web/
     php app/console c:c
     php app/console c:w
